requirements.txt
jinja2 >= 2.11.2
django-environ >= 0.4.5
cryptography >= 2.9.2
